Fix last merge (revisions 5316:5335 from trunk to ogl-es). Sorry, accidentally only the android sub-folder got checked-in and the rest was missing.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/branches/ogl-es@5366 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ include/IReadFile.h
@@ -13,7 +13,7 @@ namespace irr
 namespace io
 {
 
-	//! Interface providing read acess to a file.
+	//! Interface providing read access to a file.
 	class IReadFile : public virtual IReferenceCounted
 	{
 	public:
@@ -21,7 +21,7 @@ namespace io
 		/** \param buffer Pointer to buffer where read bytes are written to.
 		\param sizeToRead Amount of bytes to read from the file.
 		\return How many bytes were read. */
-		virtual s32 read(void* buffer, u32 sizeToRead) = 0;
+		virtual size_t read(void* buffer, size_t sizeToRead) = 0;
 
 		//! Changes position in file
 		/** \param finalPos Destination position in the file.
