+++ source/Irrlicht/CMemoryFile.h
@@ -29,7 +29,7 @@ namespace io
 		virtual ~CMemoryReadFile();
 
 		//! returns how much was read
-		virtual s32 read(void* buffer, u32 sizeToRead) _IRR_OVERRIDE_;
+		virtual size_t read(void* buffer, size_t sizeToRead) _IRR_OVERRIDE_;
 
 		//! changes position in file, returns true if successful
 		virtual bool seek(long finalPos, bool relativeMovement = false) _IRR_OVERRIDE_;
@@ -66,7 +66,7 @@ namespace io
 		virtual ~CMemoryWriteFile();
 
 		//! returns how much was written
-		virtual s32 write(const void* buffer, u32 sizeToWrite) _IRR_OVERRIDE_;
+		virtual size_t write(const void* buffer, size_t sizeToWrite) _IRR_OVERRIDE_;
 
 		//! changes position in file, returns true if successful
 		virtual bool seek(long finalPos, bool relativeMovement = false) _IRR_OVERRIDE_;
