+++ source/Irrlicht/CReadFile.h
@@ -27,7 +27,7 @@ namespace io
 		virtual ~CReadFile();
 
 		//! returns how much was read
-		virtual s32 read(void* buffer, u32 sizeToRead) _IRR_OVERRIDE_;
+		virtual size_t read(void* buffer, size_t sizeToRead) _IRR_OVERRIDE_;
 
 		//! changes position in file, returns true if successful
 		virtual bool seek(long finalPos, bool relativeMovement = false) _IRR_OVERRIDE_;
