+++ source/Irrlicht/CWriteFile.cpp
@@ -41,12 +41,12 @@ inline bool CWriteFile::isOpen() const
 
 
 //! returns how much was read
-s32 CWriteFile::write(const void* buffer, u32 sizeToWrite)
+size_t CWriteFile::write(const void* buffer, size_t sizeToWrite)
 {
 	if (!isOpen())
 		return 0;
 
-	return (s32)fwrite(buffer, 1, sizeToWrite, File);
+	return fwrite(buffer, 1, sizeToWrite, File);
 }
