+++ tools/GUIEditor/CMemoryReadWriteFile.h
@@ -16,7 +16,7 @@ namespace irr
 namespace io
 {
 
-	//! Provides write acess to an array as if it is a file.
+	//! Provides write access to an array as if it is a file.
 	class CMemoryReadWriteFile : public virtual IWriteFile, public virtual IReadFile
 	{
 	public:
@@ -25,9 +25,9 @@ namespace io
 
 		//! Reads an amount of bytes from the file.
 		//! \param buffer: Pointer to buffer of bytes to write.
-		//! \param sizeToWrite: Amount of bytes to wrtie to the file.
+		//! \param sizeToWrite: Amount of bytes to written to the file.
 		//! \return Returns how much bytes were written.
-		virtual s32 write(const void* buffer, u32 sizeToWrite);
+		virtual size_t write(const void* buffer, size_t sizeToWrite);
 
 		//! Changes position in file, returns true if successful.
 		//! \param finalPos: Destination position in the file.
@@ -45,7 +45,7 @@ namespace io
 		//! \param buffer: Pointer to buffer where to read bytes will be written to.
 		//! \param sizeToRead: Amount of bytes to read from the file.
 		//! \return Returns how much bytes were read.
-		virtual s32 read(void* buffer, u32 sizeToRead);
+		virtual size_t read(void* buffer, size_t sizeToRead);
 
 		//! Returns the current position in the file.
 		//! \return Returns the current position in the file in bytes.
